Tests for local db and garbage collector

+++ parsec/core/local_db.py
@@ -19,9 +19,9 @@ class LocalDB:
     def __init__(self, path):
         self._path = Path(path)
         self._path.mkdir(parents=True, exist_ok=True)
-        self._cache = self._path / 'cache'
+        self._cache = self._path / "cache"
         self._cache.mkdir(parents=True, exist_ok=True)
-        self._placeholders = self._path / 'placeholders'
+        self._placeholders = self._path / "placeholders"
         self._placeholders.mkdir(parents=True, exist_ok=True)
 
         # TODO: fix recursive import
@@ -34,44 +34,54 @@ class LocalDB:
     def path(self):
         return str(self._path)
 
-    def find(self, access):
-        file = self._cache / access["id"]
+    @property
+    def cache_size(self):
+        cache = str(self._cache)
+        return sum(
+            os.path.getsize(os.path.join(cache, f))
+            for f in os.listdir(cache)
+            if os.path.isfile(os.path.join(cache, f))
+        )
+
+    def _find(self, access):
         try:
-            return file
-        except FileNotFoundError:
-            pass
-        file = self._placeholders / access["id"]
-        try:
-            return file
-        except FileNotFoundError:
+            return next(
+                (
+                    directory / str(access["id"])
+                    for directory in [self._cache, self._placeholders]
+                    if (directory / str(access["id"])).exists()
+                )
+            )
+        except StopIteration:
             pass
 
     def get(self, access):
-        file = self.find(access)
-        try:
-            ciphered = file.read_bytes()
-        except FileNotFoundError:
+        file = self._find(access)
+        if not file:
             raise LocalDBMissingEntry(access)
+        ciphered = file.read_bytes()
         return self._decrypt_with_symkey(access["key"], ciphered)
 
     def set(self, access, raw: bytes, deletable=True):
-        cache = str(self._cache)
-        cache_size = sum(os.path.getsize(os.path.join(cache, f)) for f in os.listdir(cache) if os.path.isfile(os.path.join(cache, f)))
-        if cache_size > 2097152:
-            self.clean()
+        if self.cache_size > 2097152:
+            self.run_garbage_collector()
         assert isinstance(raw, (bytes, bytearray))
         ciphered = self._encrypt_with_symkey(access["key"], raw)
-        self.clear(access)
-        file = self._cache / access["id"] if deletable else self._placeholders / access["id"]
+        try:
+            self.clear(access)
+        except LocalDBMissingEntry:
+            pass
+        file = (
+            self._cache / str(access["id"]) if deletable else self._placeholders / str(access["id"])
+        )
         file.write_bytes(ciphered)
 
     def clear(self, access):
-        file = self.find(access)
-        try:
-            file.unlink()
-        except FileNotFoundError:
-            pass
+        file = self._find(access)
+        if not file:
+            raise LocalDBMissingEntry(access)
+        file.unlink()
 
-    def clean(self):
+    def run_garbage_collector(self):
         rmtree(str(self._cache))
         self._cache.mkdir(parents=True, exist_ok=True)
